Convert back to uvicorn, clear locks/setup scheduler in startServer.py
