+++ .dockerignore
@@ -25,8 +25,7 @@ jbrowse/jbrowse/website
 !main.py
 !Slurm
 !PeakLearnerSlurm.cfg
-!gunicorn.cfg.py
-!startServer.sh
+!startServer.py
 
 
 **/__pycache__
