--- startServer.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-gunicorn -c gunicorn.cfg.py main:app
